requirements.txt
supabase
firebase-admin
python-dotenv
uvicorn 